Add Dockerfile and gitlab pipeline
The dockerfile build backerei and add it to the tezos mainnet container.
The resulting container is an autonomous payout entity able to connect
to a tezos node with rpc to send payouts.

To trigger ci/cd and validate correct build on every commit, set up
a gitlab mirror of the github repo. This is currently the most
straightforward way to get a free pipeline for an open source project.

Note: modified stack.yaml to clone from https instead of git, because
otherwise a keypair is needed locally, which is not the case during
CI builds.