Add preserved cycles for alphanet and add payout delay option
On Alphanet, the rights are calculated 3 cycles in the future.

Current version of the backerei software crashes against alphanet, because
it tries to query the future baking and endorsing rights up to 5 cycles
in the future. The error encountered is "unkown seed".

At least one other person had the problem:

https://github.com/cryptiumlabs/backerei/issues/24

The PR fixes it by making "preserved_cycles" a parameter.

Also add to README to explain that param should be changed for alphanet.

Add payout delay option.

Payout delay lets you pay your delgates later (or earlier) than when
the rewards are unfrozen.