Fix issue with payout signatures in containerized case
We have previously introduced a new argument to payouts
--no-password. When that argument is passed, tezos-client
is spawned in a subprocess instead of a tty.

When running as a subprocess, the stdin/stdout handling
is different. This change fixes the subprocess case, and it does
not change the tty spawning case used by Cryptium.

This change was erroneously left out in a previous PR, so the
--no-password argument is broken unless this code gets merged.