stopgap solution for smart contract payouts
New Defi smart contracts such as Kolibri have expensive payout
operations. Backerei fee/storage/gas parameters for payout operations
are hardcoded in values that are too low for payouts to work.

This results in "simulation failure" whenever someone delegates
a Kolibri smart contract to your baker.

This PR fixes it in a very crude way, by setting high limits and fees
to every smart contract, and the usual low fee to tz addresses.

It will probably fail when there are too many Kolibri delegators
attached to your baker. Backerei has a mechanism to split payouts
across several blocks, but it does not take this new change into
account.

This should be replaced by a mechanism to estimate the fee for
every payout and charge it back to the delegator.

I don't think anyone will take that on, so the actual recommendation
is to switch to maintained payout software such as TRD.

May I have merge permissions for this repository as well?